Add support for marking rubrics
Marking rubrics are criteria which are used to form overall speaker
scores for speeches. These criteria are weighted by having different
extrema values. This is mostly used in pre-university debate, such as
WSDC or CNDC (Canada).

For the purpose, speech/ballot elements can have score subelements with
the criterion's name as attribute and score as content. There is no
extra information on the criteria so keys are not used.