Add min/maxOccurs to elements
As the default for both is 1, it wouldn't allow for specifying multiple
elements (rounds, teams, etc.). Also added some limits such as a min of
2 and max of 4 for sides, and the size of teams.

For indentation/repetition sake, both ballot elements are now merged in
an abstract element which is referenced in their old positions.

Further a new "person"-type was created for adjudicators and speakers,
using extensions for these types/elements.