Add element for interventions in speeches
An intervention is a generalization on events such as points of
information (POIs), clarifications, and privilege. More types of these
may be added in the enumeration in the future.

Each intervention has a request, which is the type of intervention as
described above, whether the request was granted, and the intervenor
(speaker) spoke or was recognized. There is not normally a content to
the tag, but can be the content to the request.