Expand use of divisions
Divisions are groupings of teams who only meet amongst themselves. With
configuration of divisions being varied, the format of the tournament
itself may change based on the division, such as having a different
style or a different number of rounds.

Such use of divisions is seen in high school formats.

To support this type of division, styles have been separated from
<tournament and made into its attributegroup (along with @language) with
references to it for <tournament and <division. Further, rounds have an
optional @division idref.