Create Tabroom XML to Archive translator
This commit adds a purpose of the repo to contain translators from
different formats, now in the 'converters/' subdirectory. The first of
these translators is for Tabroom, a tournament management system by the
National Speech & Debate Association, an American high school debate
association.

This format is used for import/export onto the Tabroom site, as well as
"Computer Assisted Tab," "Joy of Tournaments," and "TRPC." The schema
used as a basis is available on GitHub:
https://github.com/TimothyGu/tabroom/blob/master/ndt/DataFiles/TourneyData.xsd
Further, some datafiles were found and used for testing.

There seemed to be two ways to encode the rounds with results; using
<ROUND tags with flat structure or hierarchic <ROUNDRESULT tags. The
latter seems to be deprecated, with its portion commented out of the
schema. It was added anyways with the former for the completeness of
the translator.