Add fields for emails and conflicts
This commit adds optional fields, one for people to have an email
address, as well as team- and adjudicator-conflict attributes for
adjudicators. This would allow DebateXML to be used to import
participants' data as a registration step. The XML can have greater use
with minimal change. These fields are optional as inclusion may not be
appropriate for archiving though.

Refs TabbycatDebate/tabbycat#1699 for the use of DebateXML to import
demo tournaments

Closes #9